requirements.txt
asgiref==3.8.1
Django==5.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==20.1.0
psycopg2==2.9.9
python-dotenv==0.19.2
setuptools==75.1.0
sqlparse==0.5.1
